:rocket: post delete ,update ,create ,get all ,get by id

+++ src/main/java/com/esgi/api_project_annuel/Domain/repository/PostRepository.java
@@ -9,9 +9,7 @@ import java.util.List;
 
 @Repository
 public interface PostRepository extends JpaRepository<Post, Integer> {
-    Post findById(long postId);
+    Post findById(int postId);
 
     List<Post> findByUser(User user);
-
-
 }
